191017|Django|account_project

+++ 06_django_advance/03_IMAGE_UPLOAD/sns/views.py
@@ -1,17 +1,21 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_GET, require_POST
+from django.contrib.auth.decorators import login_required
 from .models import Posting, Comment
 from .forms import PostingModelForm, CommentModelForm
 
 
+# login 이 x 면 => 무조건 /account/login/
 @require_GET
 def posting_list(request):
+    # nickname = request.COOKIES.get('nickname')
     postings = Posting.objects.all()
     return render(request, 'sns/posting_list.html', {
         'postings': postings,
+        # 'nickname': nickname,
     })
 
-
+@login_required
 @require_GET
 def posting_detail(request, posting_id):
     posting = get_object_or_404(Posting, id=posting_id)
@@ -31,24 +35,27 @@ def posting_detail(request, posting_id):
 #     posting.save()
 #     return redirect(posting)  # redirect('sns:posting_detail', posting.id)
 
-
+@login_required
 @require_POST
 def create_posting(request):
     form = PostingModelForm(request.POST, request.FILES)  # 검증 & 저장 준비
     if form.is_valid():  # 검증!
-        posting = form.save()  # 저장 => Posting 객체 return
+        posting = form.save(commit=False)  # 저장 => Posting 객체 return
+        posting.user = request.user
+        posting.save()
         return redirect(posting)  # 성공하면 detail page
     else:
         return redirect('sns:posting_list')  # 실패하면 list page
 
-
+@login_required
 @require_POST
 def delete_posting(request, posting_id):
     posting = get_object_or_404(Posting, id=posting_id)
-    posting.delete()
+    if request.user == posting.user:
+        posting.delete()
     return redirect('sns:posting_list')
 
-
+@login_required
 @require_POST
 def create_comment(request, posting_id):
     posting = get_object_or_404(Posting, id=posting_id)
@@ -56,10 +63,11 @@ def create_comment(request, posting_id):
     if form.is_valid():  # content 만 값을 검증
         comment = form.save(commit=False)  # 아직 posting_id 가 비어있기 때문에, 저장하는 척 만 하고 Comment 객체 return
         comment.posting = posting
+        comment.user = request.user
         comment.save()
     return redirect(posting)
 
-
+@login_required
 @require_POST
 def delete_comment(request, comment_id, posting_id):
     posting = get_object_or_404(Posting, id=posting_id)
